refactor: shared archiveAndCloseApplication helper (v3.1.10 part 4/7)
Extract the archive flow duplicated between events/application/close.ts
and utils/api/handlers/applicationHandlers.ts into a new shared helper
at utils/application/closeWorkflow.ts::archiveAndCloseApplication().
Models on the existing utils/ticket/closeWorkflow.ts::archiveAndCloseTicket().

Both call sites collapse to a single call:
- events/application/close.ts: 130 LOC of inline archive logic →
  one helper call + result branching
- api/handlers/applicationHandlers.ts: ~110 LOC of duplicated transcript
  / forum / channel-delete code → one helper call + audit log

Helper returns `{success, archived, transcriptFailed?}` so callers can
distinguish: (a) transcript fetch died (event path replies to user with
the transcript-create error), (b) forum post failed but channel deleted
(both paths log warning and continue), (c) success.

Behavior unchanged. The API handler now returns `{success, archived}`
on archive vs. previously always `{success: true}` — `archived: false`
was previously only set on the channel-not-found short-circuit. Strictly
more informative without breaking the existing dashboard contract.

Tests still pass (1071). Build + biome clean.
