+++ src/utils/api/handlers/applicationHandlers.ts
@@ -1,12 +1,8 @@
-import type { Client, ForumChannel, ForumThreadChannel, GuildTextBasedChannel } from 'discord.js';
+import type { Client, GuildTextBasedChannel } from 'discord.js';
 import { Application } from '../../../typeorm/entities/application/Application';
-import { ArchivedApplication } from '../../../typeorm/entities/application/ArchivedApplication';
 import { ArchivedApplicationConfig } from '../../../typeorm/entities/application/ArchivedApplicationConfig';
+import { archiveAndCloseApplication } from '../../application/closeWorkflow';
 import { lazyRepo } from '../../database/lazyRepo';
-import { verifiedChannelDelete } from '../../discord/verifiedDelete';
-import { fetchMessagesAsTranscript } from '../../fetchAllMessages';
-import { enhancedLogger, LogCategory } from '../../monitoring/enhancedLogger';
-import { buildTranscript, type TicketMetadata, type TranscriptMessage } from '../../ticket/transcriptBuilder';
 import { ApiError } from '../apiError';
 import { optionalString, requireId, requireString } from '../helpers';
 import type { RouteHandler } from '../router';
@@ -14,7 +10,6 @@ import { writeAuditLog } from './auditHelper';
 
 const applicationRepo = lazyRepo(Application);
 const archivedAppConfigRepo = lazyRepo(ArchivedApplicationConfig);
-const archivedAppRepo = lazyRepo(ArchivedApplication);
 
 export function registerApplicationHandlers(client: Client, routes: Map<string, RouteHandler>): void {
   // POST /internal/guilds/:guildId/applications/:id/approve
@@ -81,83 +76,18 @@ export function registerApplicationHandlers(client: Client, routes: Map<string, 
       return { success: true, archived: false };
     }
 
-    // Build the transcript directly as Discord-markdown chunks — no files.
-    let transcriptMessages: TranscriptMessage[];
-    try {
-      transcriptMessages = await fetchMessagesAsTranscript(channel as GuildTextBasedChannel, client.user?.id ?? '');
-    } catch (error) {
-      enhancedLogger.error('Failed to fetch application messages for transcript', error as Error, LogCategory.SYSTEM, {
-        guildId,
-        channelId: app.channelId,
-      });
-      return { success: true, archived: false };
-    }
-
-    const creatorUser = await client.users.fetch(app.createdBy).catch(() => null);
-    const metadata: TicketMetadata = {
-      title: `Application: ${creatorUser?.username || 'Unknown'}`,
-      type: 'Application',
-      createdByUsername: creatorUser?.username || 'Unknown',
-      openedAt: 'createdAt' in channel && channel.createdAt instanceof Date ? channel.createdAt : new Date(),
-      closedAt: new Date(),
-      assignedToUsername: null,
-    };
-    const transcript = buildTranscript(transcriptMessages, metadata);
-
-    try {
-      const forumChannel = (await client.channels.fetch(archivedConfig.channelId)) as ForumChannel;
-
-      const existingArchive = await archivedAppRepo.findOneBy({
-        createdBy: app.createdBy,
-        guildId,
-      });
-
-      if (!existingArchive) {
-        const newPost = await forumChannel.threads.create({
-          name: creatorUser?.username || 'Unknown',
-          message: { content: transcript.header },
-        });
-        for (const chunk of transcript.chunks) {
-          await newPost.send({ content: chunk });
-        }
-        await archivedAppRepo.save(
-          archivedAppRepo.create({
-            guildId,
-            createdBy: app.createdBy,
-            messageId: newPost.id,
-          }),
-        );
-      } else if (existingArchive.messageId) {
-        const post = (await forumChannel.threads.fetch(existingArchive.messageId)) as ForumThreadChannel;
-        const separator = '\n━━━━━━━━━━━━━━━━━━━━━━━━\n';
-        await post.send({ content: separator + transcript.header });
-        for (const chunk of transcript.chunks) {
-          await post.send({ content: chunk });
-        }
-      }
-    } catch (error) {
-      enhancedLogger.error('Failed to post application transcript to forum', error as Error, LogCategory.SYSTEM, {
-        guildId,
-        channelId: app.channelId,
-      });
-    }
-
-    // Delete application channel (verified)
-    const deleteResult = await verifiedChannelDelete(channel as GuildTextBasedChannel, {
+    const result = await archiveAndCloseApplication(
+      client,
+      app,
       guildId,
-      label: 'application channel (API)',
-    });
-    if (!deleteResult.success) {
-      enhancedLogger.error('Application channel persisted after API archive', undefined, LogCategory.ERROR, {
-        guildId,
-        channelId: app.channelId,
-      });
-    }
+      channel as GuildTextBasedChannel,
+      archivedConfig.channelId,
+    );
 
     const triggeredBy = optionalString(body, 'triggeredBy');
     await writeAuditLog(guildId, 'application.archive', triggeredBy, {
       applicationId: app.id,
     });
-    return { success: true };
+    return { success: result.success, archived: result.archived };
   });
 }
